requirements.txt
boto3==1.40.53
chromadb==1.1.1
pypdf==6.1.1
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.50.0
